requirements.txt
python-telegram-bot==21.6
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/refs/heads/master.tar.gz
bgutil-ytdlp-pot-provider
